requirements.txt
# === Flask Core ===
Flask>=2.3.0

# === Formularios y Validaciones ===
Flask-WTF>=1.1.1
WTForms>=3.1.2
email-validator>=1.3.1

# === Base de Datos y ORM ===
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.4
PyMySQL>=1.1.0

# === Variables de entorno (opcional pero útil) ===
python-dotenv>=1.0.1

# === Autenticación y Seguridad (opcional si usarás login/register) ===
Flask-Login>=0.6.3
Flask-Bcrypt>=1.0.1
